+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version [0.22.0](https://github.com/cedx/base/compare/v0.21.1...v0.22.0)
+- Added the `result` property to the `UI.Components.DialogBox` component.
 - Fixed the `css()` and `html()` tag functions to handle arrays.
 - Removed the `UI.Components.DialogButton` component.
 - Renamed the `UI.Components.IMessage` interface to `IDialogMessage`.